+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { AuthProvider } from '@contexts/AuthContext';
 import { ModalProvider } from '@contexts/ModalContext';
 import styled from '@emotion/styled';
-import Router from '@shared/Router';
+import Router from '@shared/routes/Router';
 
 const Container = styled.div(({ theme }) => ({
   backgroundColor: theme.colors.semantic.backgroundDefault,
